Rename document view to split pane view
