+++ src/document/SplitPaneView.css
